@@ -1,13 +1,12 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
+import type { NextConfig } from 'next'
+
+const nextConfig: NextConfig = {
     reactStrictMode: true,
     images: {
-        domains: ['fakestoreapi.com', 'localhost'],
         remotePatterns: [
             {
                 protocol: 'https',
                 hostname: 'fakestoreapi.com',
-                port: '',
                 pathname: '/img/**',
             },
         ],
@@ -15,6 +14,14 @@ const nextConfig = {
     env: {
         NEXT_PUBLIC_API_URL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:8000',
     },
+    eslint: {
+        // Ensure lint warnings are treated as errors in builds
+        ignoreDuringBuilds: false,
+    },
+    typescript: {
+        // Ensure type errors are treated as errors in builds
+        ignoreBuildErrors: false,
+    },
 }
 
-module.exports = nextConfig
+export default nextConfig
