@@ -1,6 +1,8 @@
-module.exports = {
+const config = {
     plugins: {
         tailwindcss: {},
         autoprefixer: {},
     },
 }
+
+export default config
